+++ public/app/features/dashboard/dashgrid/PanelHeader/PanelHeaderCorner.tsx
@@ -4,7 +4,7 @@ import { Tooltip } from '@grafana/ui';
 import { PanelModel } from 'app/features/dashboard/panel_model';
 import templateSrv from 'app/features/templating/template_srv';
 import { LinkSrv } from 'app/features/panel/panellinks/link_srv';
-import { getTimeSrv, TimeSrv } from 'app/features/dashboard/time_srv';
+import { getTimeSrv, TimeSrv } from 'app/features/dashboard/services/TimeSrv';
 
 enum InfoModes {
   Error = 'Error',
